update sklearn dataclasses
