+++ src/tabular_trees/__init__.py
@@ -8,8 +8,8 @@ from .explain.shapley_values import calculate_shapley_values
 from .lightgbm.booster_string import BoosterString
 from .lightgbm.editable_booster import EditableBooster
 from .lightgbm.lightgbm_tabular_trees import LightGBMTabularTrees
-from .sklearn.scikit_learn_hist_tabular_trees import ScikitLearnHistTabularTrees
-from .sklearn.scikit_learn_tabular_trees import ScikitLearnTabularTrees
+from .sklearn.sklearn_hist_tabular_trees import ScikitLearnHistTabularTrees
+from .sklearn.sklearn_tabular_trees import ScikitLearnTabularTrees
 from .trees import TabularTrees, export_tree_data
 from .validate import MonotonicConstraintResults, validate_monotonic_constraints
 from .xgboost.dump_parser import XGBoostParser
@@ -27,7 +27,7 @@ with contextlib.suppress(ImportError):
     from .lightgbm import lightgbm_tabular_trees
 
 with contextlib.suppress(ImportError):
-    from .sklearn import scikit_learn_tabular_trees
+    from .sklearn import sklearn_tabular_trees
 
 with contextlib.suppress(ImportError):
     from .xgboost import xgboost_tabular_trees
